Move wsgi.py to avalonstar/.
